requirements.txt
google-generativeai
faiss-cpu
numpy
pdfplumber
streamlit
python-dotenv
sentence-transformers